updateCodex/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.0
python-dotenv>=1.0
PyMySQL>=1.0
Flask-Mail>=0.9
passlib[bcrypt]>=1.7
Flask-Cors>=3.0
